fix bit-start
