+++ e2e/functionalities/repository-hooks.e2e.ts
@@ -47,7 +47,7 @@ describe('repository-hooks', function () {
       it('should run the on read hook', () => {
         const regex = new RegExp('on read run', 'g');
         const count = importOutput.match(regex);
-        // total 3 objects - component, version and file
+        // total 4 objects - component, version, version-history and file
         // the reason for the 5 reading is that it happens in two places.
         // 1. repository.load(), it reads 2 files, the component and the version objects.
         // 2. repository.loadRaw(), it reads 3 files, component, version and file.
@@ -56,7 +56,7 @@ describe('repository-hooks', function () {
 
         // if this test fails, and the number is bigger than 5. this could indicate a serious performance issue,
         // because for components with large amount of versions, this could become a huge number.
-        expect(count).to.have.lengthOf(5);
+        expect(count).to.have.lengthOf(6);
       });
       it('should be able to import the component as usual', () => {
         expect(importOutput).to.have.string('successfully imported one component');
